AI-Based Substation Monitoring System for Predictive Maintenance/requirements.txt
fastapi 
uvicorn 
kafka-python
cassandra-driver 
scikit-learn 
joblib 
numpy
requests
